Split monitor test into tests that need and don't need mocking
